ivp5: store D8F i8 scales as half

+++ metal/d8f.metal
@@ -1097,8 +1097,8 @@ kernel void d8f_down_lut_score_i8_selected_batch(
       mid + ulong(token) * ulong(args.mid_token_stride) +
       ulong(slot) * ulong(args.mid_slot_stride);
   const device char *cb = i8_codebook + ulong(i8.offset) + ulong(code) * 8ul;
-  const device float *i8_scales = (const device float *)(i8_codebook + ulong(i8.scale_offset));
-  const float scale = i8_scales[code];
+  const device half *i8_scales = (const device half *)(i8_codebook + ulong(i8.scale_offset));
+  const float scale = float(i8_scales[code]);
   if (scale == 0.0f) {
     const device half *hcb = (const device half *)(pack + rec.codebook_offset + ulong(code) * 16ul);
     score[out_index] =
@@ -1853,7 +1853,7 @@ kernel void d8f_down_lut_direct_codes_i8_selected_batch_tile16_cbsram2048(
   device const char *i8_codebook    [[buffer(10)]],
   threadgroup float *partial        [[threadgroup(0)]],
   threadgroup char *cb_cache        [[threadgroup(1)]],
-  threadgroup float *scale_cache    [[threadgroup(2)]],
+  threadgroup half *scale_cache     [[threadgroup(2)]],
   uint tid [[thread_index_in_threadgroup]],
   ushort tiisg [[thread_index_in_simdgroup]],
   ushort sgitg [[simdgroup_index_in_threadgroup]],
@@ -1875,7 +1875,7 @@ kernel void d8f_down_lut_direct_codes_i8_selected_batch_tile16_cbsram2048(
     const bool use_i8_cache = i8.k == rec.k && i8.k > 0u && i8.k <= cache_k_cap;
     if (use_i8_cache) {
       const device char *i8_cb = i8_codebook + ulong(i8.offset);
-      const device float *i8_scales = (const device float *)(i8_codebook + ulong(i8.scale_offset));
+      const device half *i8_scales = (const device half *)(i8_codebook + ulong(i8.scale_offset));
       const uint cb_vals = i8.k << 3;
       for (uint ci = tid; ci < cb_vals; ci += 256u) cb_cache[ci] = i8_cb[ci];
       for (uint ci = tid; ci < i8.k; ci += 256u) {
@@ -1897,7 +1897,7 @@ kernel void d8f_down_lut_direct_codes_i8_selected_batch_tile16_cbsram2048(
           if (row >= args.rows) continue;
           const uint code = uint(codes[(ulong(slot) * ulong(args.rows) + ulong(row)) * ulong(groups) + ulong(group)]);
           if (code >= rec.k) continue;
-          const float scale = code < i8.k ? scale_cache[code] : 0.0f;
+          const float scale = code < i8.k ? float(scale_cache[code]) : 0.0f;
           if (scale != 0.0f) {
             threadgroup const char *cb = cb_cache + ulong(code) * 8ul;
             acc[rr] += scale * (
@@ -2629,7 +2629,7 @@ kernel void d8f_down_sum_selected_weighted_batch_tile16_recbuf_native_codes_i8_c
   device const char *i8_codebook         [[buffer(9)]],
   threadgroup float *partial             [[threadgroup(0)]],
   threadgroup char *cb_cache             [[threadgroup(1)]],
-  threadgroup float *scale_cache         [[threadgroup(2)]],
+  threadgroup half *scale_cache          [[threadgroup(2)]],
   uint tid [[thread_index_in_threadgroup]],
   ushort tiisg [[thread_index_in_simdgroup]],
   ushort sgitg [[simdgroup_index_in_threadgroup]],
@@ -2667,7 +2667,7 @@ kernel void d8f_down_sum_selected_weighted_batch_tile16_recbuf_native_codes_i8_c
     const bool use_i8_cache = i8.k == rec.k && i8.k > 0u && i8.k <= 2048u;
     if (use_i8_cache) {
       const device char *i8_cb = i8_codebook + ulong(i8.offset);
-      const device float *i8_scales = (const device float *)(i8_codebook + ulong(i8.scale_offset));
+      const device half *i8_scales = (const device half *)(i8_codebook + ulong(i8.scale_offset));
       const uint cb_vals = i8.k << 3;
       for (uint ci = tid; ci < cb_vals; ci += 256u) cb_cache[ci] = i8_cb[ci];
       for (uint ci = tid; ci < i8.k; ci += 256u) {
@@ -2709,7 +2709,7 @@ kernel void d8f_down_sum_selected_weighted_batch_tile16_recbuf_native_codes_i8_c
           code = (w >> shift) & rec.mask;
         }
         if (code >= rec.k) continue;
-        const float scale = (use_i8_cache && code < i8.k) ? scale_cache[code] : 0.0f;
+        const float scale = (use_i8_cache && code < i8.k) ? float(scale_cache[code]) : 0.0f;
         if (scale != 0.0f) {
           threadgroup const char *cb = cb_cache + ulong(code) * 8ul;
           acc[rr] += rw * scale * (
